Start of new studentsessions
