+++ Nexpo/Repositories/StudentSessionApplicationRepository.cs
@@ -10,7 +10,7 @@ namespace Nexpo.Repositories
     public interface IStudentSessionApplicationRepository
     {
         public Task<bool> ApplicationExists(int studentId, int companyId);
-        public Task<IEnumerable<StudentSessionApplication>> GetAllUnassigned();
+        public Task<IEnumerable<StudentSessionApplication>> GetByCompanyAndStudent(int studentId, int companyId);
         public Task<IEnumerable<StudentSessionApplication>> GetAllForStudent(int studentId);
         public Task<IEnumerable<StudentSessionApplication>> GetAllForCompany(int companyId);
         public Task<StudentSessionApplication> Get(int id);
@@ -33,6 +33,11 @@ namespace Nexpo.Repositories
             return await _context.StudentSessionApplications.AnyAsync(a => a.StudentId == studentId && a.CompanyId == companyId);
         }
 
+        public async Task<IEnumerable<StudentSessionApplication>> GetByCompanyAndStudent(int studentId, int companyId)
+        {
+            return await _context.StudentSessionApplications.Where(a => a.StudentId == studentId && a.CompanyId == companyId).ToListAsync();
+        }
+
         public async Task<IEnumerable<StudentSessionApplication>> GetAllForStudent(int studentId)
         {
             return await _context.StudentSessionApplications.Where(a => a.StudentId == studentId).ToListAsync();
@@ -41,11 +46,6 @@ namespace Nexpo.Repositories
         public async Task<IEnumerable<StudentSessionApplication>> GetAllForCompany(int companyId)
         {
             return await _context.StudentSessionApplications.Where(a => a.CompanyId == companyId).ToListAsync();
-        }
-
-        public async Task<IEnumerable<StudentSessionApplication>> GetAllUnassigned()
-        {
-            return await _context.StudentSessionApplications.Where(a => a.StudentSession == null).ToListAsync();
         }
         
         public async Task<StudentSessionApplication> Get(int id)
